requirements.txt
django-mptt-admin==2.4.1
django-mptt==0.14.0
Django==4.1.6
Pillow==9.4.0
django-ckeditor==6.5.1